removed unnecessary files from actions folder
