+++ app/repositories/user.js
@@ -1,18 +1,11 @@
-import Realm from 'realm';
-import {
-  Inventory,
-  Species,
-  Polygons,
-  Coordinates,
-  OfflineMaps,
-  User,
-  AddSpecies,
-} from '../actions/Schemas';
-import { bugsnag } from '../utils';
-
-import { getUserInformationFromServer } from '../actions/User';
+import axios from 'axios';
 import Auth0 from 'react-native-auth0';
 import Config from 'react-native-config';
+import Realm from 'realm';
+import { APIConfig } from '../actions/Config';
+import { AddSpecies, Coordinates, Inventory, OfflineMaps, Polygons, Species, User } from './schema';
+import { bugsnag } from '../utils';
+import getSessionData from '../utils/sessionId';
 
 // AUTH0 CONFIG
 const auth0 = new Auth0({ domain: Config.AUTH0_DOMAIN, clientId: Config.AUTH0_CLIENT_ID });
@@ -130,3 +123,72 @@ export const LoginDetails = () => {
 };
 
 //  ---------------- AUTH0 ACTIONS END----------------
+
+export const getUserInformation = () => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    }).then((realm) => {
+      const User = realm.objectForPrimaryKey('User', 'id0001');
+      if (User) {
+        resolve({
+          email: User.email,
+          firstName: User.firstname,
+          lastName: User.lastname,
+          country: User.country,
+        });
+      } else {
+        resolve({ email: '', firstName: '', lastName: '' });
+      }
+    });
+  });
+};
+
+export const getUserInformationFromServer = (navigation) => {
+  const { protocol, url } = APIConfig;
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    }).then((realm) => {
+      const User = realm.objectForPrimaryKey('User', 'id0001');
+      let userToken = User.accessToken;
+      console.log(userToken, 'Token');
+      getSessionData().then((sessionData) => {
+        axios({
+          method: 'GET',
+          url: `${protocol}://${url}/app/profile`,
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `OAuth ${userToken}`,
+            'x-session-id': sessionData,
+          },
+        })
+          .then((data) => {
+            realm.write(() => {
+              const { email, firstname, lastname, country } = data.data;
+              realm.create(
+                'User',
+                {
+                  id: 'id0001',
+                  email,
+                  firstname,
+                  lastname,
+                  country,
+                },
+                'modified',
+              );
+            });
+            resolve(data.data);
+          })
+          .catch((err) => {
+            console.error('err.response.status =>>', err);
+            if (err.response.status === 303) {
+              navigation.navigate('SignUp');
+            }
+            reject(err);
+          });
+      });
+      // realm.close();
+    });
+  });
+};
